requirements.txt
langchain==0.3.14
langchain-openai==0.2.14
langchain-core==0.3.29
langchain-community==0.3.14
tavily-python==0.5.0
requests==2.32.3
python-dotenv==1.0.1
openai>=1.0.0
pydantic>=2.0.0